Not my fault that Anope does not understand "role:hostmask" notation
Anope shall handle this, this is not my fault that Anope
blindly strips hostname with '@' if it sees ':' in nickname part.

I also don't want to turn it into extban notation since, well,
IRC protocol permits me doing so. There are alot of things Anope
still does not understand, for example, IPv6 address with CIDR,
and it is not my fault if this pile of C++ ugly code cannot handle this.
And since this ircd can work without services at all, any services
package can be used with it. /rant

Still, accept '#' as a fallback until I will find out why Anope does that.